requirements.txt
tqdm==4.66.5
openai==1.43.0
elasticsearch==8.15.0
datasets==2.21.0
python-dotenv==1.0.1
black==24.8.0
isort==5.13.2